refactor: create utils backend module
