+++ scripts/seedZeroHashes.ts
@@ -1,12 +1,12 @@
-import { dbConnect, dbDisconnect } from "src/utils/backend/database"
+import { connectDatabase, disconnectDatabase } from "src/utils/backend/database"
 import { seedZeroHashes } from "src/utils/backend/seeding"
 
 async function main() {
-    await dbConnect()
+    await connectDatabase()
 
     await seedZeroHashes()
 
-    await dbDisconnect()
+    await disconnectDatabase()
 }
 
 main()
