+++ src/controllers/groups/getGroup.ts
@@ -1,8 +1,8 @@
 import { NextApiRequest, NextApiResponse } from "next"
 import { checkGroup, getGroup } from "src/core/groups"
 import { Provider } from "src/types/groups"
-import { dbConnect } from "src/utils/backend/database"
-import logger from "src/utils/backend/logger"
+import { logger } from "src/utils/backend"
+import { connectDatabase } from "src/utils/backend/database"
 
 export default async function getGroupController(req: NextApiRequest, res: NextApiResponse) {
     if (req.method !== "GET") {
@@ -24,7 +24,7 @@ export default async function getGroupController(req: NextApiRequest, res: NextA
     }
 
     try {
-        await dbConnect()
+        await connectDatabase()
 
         const group = await getGroup(provider as Provider, name as any)
