+++ src/controllers/groups/handlePoapIdentityCommitment.ts
@@ -3,8 +3,8 @@ import { NextApiRequest, NextApiResponse } from "next"
 import { appendLeaf, deleteLeaf } from "src/core/groups/mts"
 import { getPoapEventsByAddress, PoapEvent } from "src/core/poap"
 import { Web3Provider } from "src/types/groups"
-import { dbConnect } from "src/utils/backend/database"
-import logger from "src/utils/backend/logger"
+import { logger } from "src/utils/backend"
+import { connectDatabase } from "src/utils/backend/database"
 
 export default async function handlePoapIdentityCommitmentController(req: NextApiRequest, res: NextApiResponse) {
     const name = req.query?.name as string
@@ -27,7 +27,7 @@ export default async function handlePoapIdentityCommitmentController(req: NextAp
             throw new Error(`The address does not hold the group POAP token`)
         }
 
-        await dbConnect()
+        await connectDatabase()
 
         if (req.method === "POST") {
             await appendLeaf(Web3Provider.POAP, name, identityCommitment)
