+++ src/controllers/groups/handleTelegramIdentityCommitment.ts
@@ -1,8 +1,8 @@
 import { TelegramUser } from "@interrep/db"
 import { NextApiRequest, NextApiResponse } from "next"
 import { appendLeaf, deleteLeaf } from "src/core/groups/mts"
-import { dbConnect } from "src/utils/backend/database"
-import logger from "src/utils/backend/logger"
+import { connectDatabase } from "src/utils/backend/database"
+import { logger } from "src/utils/backend"
 import { sha256 } from "src/utils/common/crypto"
 
 export default async function handleTelegramIdentityCommitmentController(req: NextApiRequest, res: NextApiResponse) {
@@ -16,7 +16,7 @@ export default async function handleTelegramIdentityCommitmentController(req: Ne
     }
 
     try {
-        await dbConnect()
+        await connectDatabase()
 
         const hashId = sha256(telegramUserId + name)
         const telegramUser = await TelegramUser.findByHashId(hashId)
