+++ src/controllers/groups/hasJoinedAGroup.ts
@@ -1,8 +1,8 @@
+import { OAuthAccount } from "@interrep/db"
 import { NextApiRequest, NextApiResponse } from "next"
 import { getSession } from "next-auth/client"
-import { OAuthAccount } from "@interrep/db"
-import { dbConnect } from "src/utils/backend/database"
-import logger from "src/utils/backend/logger"
+import { logger } from "src/utils/backend"
+import { connectDatabase } from "src/utils/backend/database"
 
 export default async function hasJoinedAGroupController(req: NextApiRequest, res: NextApiResponse) {
     if (req.method !== "GET") {
@@ -18,7 +18,7 @@ export default async function hasJoinedAGroupController(req: NextApiRequest, res
     }
 
     try {
-        await dbConnect()
+        await connectDatabase()
 
         const account = await OAuthAccount.findById(session.accountId)
