+++ src/controllers/trees/getLeaves.ts
@@ -1,7 +1,7 @@
 import { MerkleTreeNode } from "@interrep/db"
 import { NextApiRequest, NextApiResponse } from "next"
-import { dbConnect } from "src/utils/backend/database"
-import logger from "src/utils/backend/logger"
+import { logger } from "src/utils/backend"
+import { connectDatabase } from "src/utils/backend/database"
 
 export default async function getLeavesController(req: NextApiRequest, res: NextApiResponse) {
     if (req.method !== "GET") {
@@ -24,7 +24,7 @@ export default async function getLeavesController(req: NextApiRequest, res: Next
     }
 
     try {
-        await dbConnect()
+        await connectDatabase()
 
         const root = await MerkleTreeNode.findOne({ hash: rootHash })
