+++ src/core/contracts/ReputationBadge/exists.ts
@@ -1,6 +1,6 @@
 import { OAuthProvider } from "@interrep/reputation"
 import { ContractName } from "src/config"
-import getBackendContractInstance from "src/utils/backend/getBackendContractInstance"
+import { getBackendContractInstance } from "src/utils/backend"
 import { getContractAddress } from "src/utils/common"
 
 export default async function exists(tokenId: string, provider: OAuthProvider): Promise<boolean> {
