+++ src/core/contracts/ReputationBadge/safeMint.ts
@@ -1,7 +1,7 @@
 import { OAuthProvider } from "@interrep/reputation"
 import { ContractReceipt } from "ethers"
 import { ContractName } from "src/config"
-import getBackendContractInstance from "src/utils/backend/getBackendContractInstance"
+import { getBackendContractInstance } from "src/utils/backend"
 import { getContractAddress } from "src/utils/common"
 
 export default async function safeMint(to: string, tokenId: string, provider: OAuthProvider): Promise<ContractReceipt> {
