+++ src/utils/backend/database/database.ts
@@ -1,8 +1,9 @@
+/* istanbul ignore file */
 import { clear, connect, disconnect, getState } from "@interrep/db"
 import config from "src/config"
-import logger from "./logger"
+import { logger } from "src/utils/backend"
 
-export async function dbConnect() {
+export async function connectDatabase() {
     if (getState() !== 0) {
         return
     }
@@ -20,7 +21,7 @@ export async function dbConnect() {
     logger.info("Mongo db has been connected")
 }
 
-export async function dbDisconnect() {
+export async function disconnectDatabase() {
     if (getState() !== 1) {
         return
     }
@@ -30,7 +31,7 @@ export async function dbDisconnect() {
     logger.info("Mongo db has been disconnected")
 }
 
-export async function dbClear() {
+export async function clearDatabase() {
     if (getState() !== 1) {
         return
     }
