+++ src/utils/backend/database/index.ts
@@ -1,0 +1,15 @@
+import { connectDatabase, clearDatabase, disconnectDatabase } from "./database"
+import {
+    connectDatabase as connectTestingDatabase,
+    disconnectDatabase as disconnectTestingDatabase,
+    clearDatabase as clearTestingDatabase
+} from "./testingDatabase"
+
+export {
+    connectDatabase,
+    clearDatabase,
+    disconnectDatabase,
+    connectTestingDatabase,
+    disconnectTestingDatabase,
+    clearTestingDatabase
+}
