+++ src/utils/backend/database/testingDatabase.ts
@@ -1,3 +1,4 @@
+/* istanbul ignore file */
 import { clear, connect, disconnect, drop } from "@interrep/db"
 import { MongoMemoryServer } from "mongodb-memory-server"
 
@@ -8,7 +9,7 @@ export async function connectDatabase() {
     await connect(mms.getUri())
 }
 
-export async function dropDatabaseAndDisconnect() {
+export async function disconnectDatabase() {
     await drop()
     await disconnect()
     await mms.stop()
