+++ src/utils/backend/getBackendContractInstance.ts
@@ -1,6 +1,6 @@
 import { Contract } from "ethers"
 import { ContractName } from "src/config"
-import getSigner from "src/utils/backend/getSigner"
+import { getSigner } from "src/utils/backend"
 import { getContractInstance } from "src/utils/common"
 
 export default async function getBackendContractInstance(
