+++ src/utils/backend/index.ts
@@ -1,0 +1,8 @@
+import getBackendContractInstance from "./getBackendContractInstance"
+import getProvider from "./getProvider"
+import getSigner from "./getSigner"
+import logger from "./logger"
+import removeDBFields from "./removeDBFields"
+import setAPIMiddleware from "./setAPIMiddleware"
+
+export { getSigner, getProvider, getBackendContractInstance, setAPIMiddleware, logger, removeDBFields }
