--- src/utils/backend/isTransactionConfirmed.ts
@@ -1,10 +1,0 @@
-import getProvider from "src/utils/backend/getProvider"
-
-export default async function isTransactionConfirmed(txHash: string): Promise<boolean> {
-    const provider = getProvider()
-    const receipt = await provider.getTransactionReceipt(txHash)
-
-    if (!receipt || receipt.confirmations < 3) return false
-
-    return true
-}
