+++ src/utils/backend/setAPIMiddleware.ts
@@ -4,7 +4,7 @@ import { NextApiRequest, NextApiResponse } from "next"
  * Helper method to wait for a middleware to execute before continuing
  * and to throw an error when an error happens in a middleware.
  */
-export default function apiMiddleware(
+export default function setAPIMiddleware(
     middleware: (req: NextApiRequest, res: NextApiResponse, next: (err?: any) => any) => void
 ) {
     return (req: NextApiRequest, res: NextApiResponse) =>
